+++ NINA.Profile/Interfaces/ITelescopeSettings.cs
@@ -23,5 +23,7 @@ namespace NINA.Profile.Interfaces {
         string SnapPortStart { get; set; }
         string SnapPortStop { get; set; }
         bool NoSync { get; set; }
+        bool PrimaryReversed { get; set; }
+        bool SecondaryReversed { get; set; }
     }
 }